Same thing for the other migration that got moved

+++ db/migrate/20170831151121_add_seasons_to_submission_scores.rb
@@ -1,15 +1,19 @@
 class AddSeasonsToSubmissionScores < ActiveRecord::Migration[5.1]
   def up
-    add_column :submission_scores,
-      :seasons,
-      :text,
-      array: true,
-      default: []
+    unless column_exists?(:submission_scores, :seasons)
+      add_column :submission_scores,
+        :seasons,
+        :text,
+        array: true,
+        default: []
 
-    SubmissionScore.update_all(seasons: [2017])
+      SubmissionScore.update_all(seasons: [2017])
+    end
   end
 
   def down
-    remove_column :submission_scores, :seasons
+    if column_exists?(:submission_scores, :seasons)
+      remove_column :submission_scores, :seasons
+    end
   end
 end
